Conflicts:
	include/mapnik/feature_type_style.hpp
	src/feature_type_style.cpp

+++ include/mapnik/svg/svg_renderer.hpp
@@ -2,7 +2,7 @@
  *
  * This file is part of Mapnik (c++ mapping toolkit)
  *
- * Copyright (C) 2010 Artem Pavlenko
+ * Copyright (C) 2011 Artem Pavlenko
  *
  * This library is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -86,7 +86,7 @@ public:
             u = 0;
         else if (u > 1)
             u = 1;
-        return u*d;
+        return static_cast<int>(u*d);
     }
 private:
     double x1_;
